Fix Edkt top level type not being used
