+++ types/index.ts
@@ -163,6 +163,17 @@ declare global {
 
 // Top level API interfaces
 
+interface Config {
+  audienceDefinitions: AudienceDefinition[];
+  pageFeatures?: Record<string, PageFeatureResult>;
+  pageMetadata?: Record<string, string | number | boolean>;
+  vendorIds?: number[];
+  omitGdprConsent?: boolean;
+  featureStorageSize?: number;
+}
+
 export interface Edkt {
-  run: () => Promise<void>;
+  run: (config: Config) => Promise<void>;
+  getMatchedAudiences: () => MatchedAudience[];
+  getCopyOfPageViews: () => PageView[];
 }
